Move Artist container, test, and scss to /containers
