+++ src/containers/Artist/Artist.js
@@ -1,9 +1,9 @@
 import React, {Component} from "react";
-import ArtistBio from "./ArtistBio/ArtistBio";
-import ArtistPerformances from "../../containers/Artists/ArtistPerformances/ArtistPerformances";
-import MoneyRaised from "./MoneyRaised/MoneyRaised";
-import SupportedCauses from "../../containers/Artists/SupportedCauses/SupportedCauses";
-import RelatedArtists from "../../containers/Artists/RelatedArtists/RelatedArtists";
+import ArtistBio from "../../components/Artist/ArtistBio/ArtistBio";
+import ArtistPerformances from "../Artists/ArtistPerformances/ArtistPerformances";
+import MoneyRaised from "../../components/Artist/MoneyRaised/MoneyRaised";
+import SupportedCauses from "../Artists/SupportedCauses/SupportedCauses";
+import RelatedArtists from "../Artists/RelatedArtists/RelatedArtists";
 
 class ArtistContainer extends Component {
   render() {
