+++ src/containers/Artist/artist.scss
