+++ src/main.scss
@@ -11,7 +11,7 @@
 // COMPONENTS
 @import "components/Header/header.scss";
 @import "components/Footer/footer.scss";
-@import "components/Artist/artist.scss";
+@import "containers/Artist/artist.scss";
 @import "containers/Artists/ArtistPerformances/artistperformances.scss";
 @import "components/Artist/ArtistPerformance/artistperformance.scss";
 @import "components/Artist/MoneyRaised/MoneyRaised.scss";
